move our rewrite of WPI's PIDController to the rewrittenWPIclasses
